requirements.txt
gtts==2.4.1
avro.py==2025.11.3